requirements.txt
autoenv==1.0.0
beautifulsoup4==4.5.3
cf-deployment-tracker==1.0.2
click==6.6
cloudant==2.4.0
Django==1.10.3
django-geoposition==0.3.0
django-modelcluster==3.1
django-taggit==0.22.0
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.2
Flask==0.11.1
gTTS==1.1.6
gTTS-token==1.1.1
html5lib==0.999999999
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.4.2
plumbum==1.6.3
pyglet==1.2.4
PyMySQL==0.7.9
pytz==2017.2
requests==2.12.2
rpyc==3.3.0
six==1.10.0
Unidecode==0.4.20
virtualenv==15.1.0
wagtail==1.9
webencodings==0.5.1
Werkzeug==0.11.11
Willow==0.4
psycopg2==2.7.3.1
djangorestframework==3.6.3
